Fixed up the backup file to latest pre-refactor state
